fix updater usage
