+++ pkg/controller/status_updater.go
@@ -34,10 +34,17 @@ const (
 	progressDebounceDuration = 3 * time.Second
 )
 
+type Progress struct {
+	MetaCompleted  uint64 `json:"meta_completed"`
+	MetaTotal      uint64 `json:"meta_total"`
+	BytesToCompact uint64 `json:"bytes_to_compact"`
+	BytesCompacted uint64 `json:"bytes_compacted"`
+}
+
 type CompactStatusUpdaterInterface interface {
 	OnStart(ctx context.Context, compact *v1alpha1.CompactBackup)
-	OnProgress(ctx context.Context, p Progress, compact *v1alpha1.CompactBackup)
-	OnFinish(ctx context.Context, compact *v1alpha1.CompactBackup)
+	OnProgress(ctx context.Context, compact *v1alpha1.CompactBackup, p Progress)
+	OnFinish(ctx context.Context, compact *v1alpha1.CompactBackup, err error)
 }
 
 type CompactStatusUpdater struct {
@@ -123,7 +130,7 @@ func (r *CompactStatusUpdater) OnProgress(ctx context.Context, p Progress, compa
 	r.UpdateStatus(compact, "", progress, "")
 }
 
-func (r *CompactStatusUpdater) OnFinish(ctx context.Context, err error, compact *v1alpha1.CompactBackup) {
+func (r *CompactStatusUpdater) OnFinish(ctx context.Context, compact *v1alpha1.CompactBackup, err error) {
 	if err != nil {
 		r.Event(compact, corev1.EventTypeWarning, "Failed", err.Error())
 		r.UpdateStatus(compact, string(v1alpha1.BackupFailed), "", err.Error())
